Moved isuserowner to SessionUtil
